+++ tests/.vscode/tasks.json
@@ -7,7 +7,7 @@
             "type": "process",
             "args": [
                 "build",
-                "${workspaceFolder}/GilGoblin.Test.csproj",
+                "${workspaceFolder}/GilGoblin.Tests.csproj",
                 "/property:GenerateFullPaths=true",
                 "/consoleloggerparameters:NoSummary"
             ],
@@ -19,7 +19,7 @@
             "type": "process",
             "args": [
                 "publish",
-                "${workspaceFolder}/GilGoblin.Test.csproj",
+                "${workspaceFolder}/GilGoblin.Tests.csproj",
                 "/property:GenerateFullPaths=true",
                 "/consoleloggerparameters:NoSummary"
             ],
@@ -33,7 +33,7 @@
                 "watch",
                 "run",
                 "--project",
-                "${workspaceFolder}/GilGoblin.Test.csproj"
+                "${workspaceFolder}/GilGoblin.Tests.csproj"
             ],
             "problemMatcher": "$msCompile"
         }
