+++ tests/web/MarketDataGatewayTest.cs
@@ -3,7 +3,7 @@ using GilGoblin.Web;
 using NSubstitute;
 using NUnit.Framework;
 
-namespace GilGoblin.Test.Web
+namespace GilGoblin.Tests.Web
 {
     [TestFixture]
     public class CraftingCalculatorTest
